requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1
feedparser==6.0.10
pyyaml==6.0.1
requests==2.31.0
sendgrid==6.11.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
pytz==2024.1
python-dateutil==2.8.2
google-genai
